data_access: add existing OAuth implementation
The code is hardly tested, and isn't all that good, but it does work and is something that we can build on later.

IntelliJ complains about calling an internal okhttp function, but I don't see any way around it. The code runs in any case. As such, suppress warnings about it.