data_access: fix Google token refreshing
We previously assumed Google would want the login code repeatedly, but
it turns out they want different information. Provide it unless we're
requesting the first token.

I'm not especially a fan of the special-casing, but it works for now,
and I don't know if it's worth adding more classes for this.

The relevant documentation is at:

https://developers.google.com/identity/protocols/oauth2/native-app#offline