general: remove IntelliJ project files
Referring to the following:

https://intellij-support.jetbrains.com/hc/en-us/articles/206544839-How-to-manage-projects-under-Version-Control-Systems

There's two types of files here:

* .iml files, which we don't need (we're a Maven-based project)

* .idea/... files, which it does recommend sharing

This removes all of them. Why? (a) It makes the project less
IDE-independent, since we have IntelliJ files in the repo; (b) they
cause annoyance when something changes for no good reason; (c) given the
Maven configuration in pom.xml, it should autogenerate anyways.

As such, remove them and add them to the gitignore. If this becomes a
problem, we can just allow them back in.